Split API tests into per-api files
