+++ ce/tests/api/test_models.py
@@ -1,0 +1,9 @@
+import pytest
+
+from ce.api import models
+
+
+def test_models(populateddb):
+    sesh = populateddb.session
+    rv = models(sesh, 'ce')
+    assert rv
